requirements.txt
python-dotenv>=1.0.0
colorama>=0.4.6
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
langdetect>=1.0.9
PyYAML>=6.0
